requirements.txt
Django==1.9.4
django-celery==3.1.17
redis==2.10.5
djangorestframework==3.3.3
